Feature/R4SU-55 initial refactoring? gameEngine and gamePainter logic divided
