+++ docs/game.md
@@ -72,11 +72,11 @@ gameEngine/types:
 
 gameEngine:
 
-    collisionManager.ts - определяет коллизии по текущему стейту
+    collisionManager.ts - определяет коллизии по текущему стейту и выставляет стейт при столкновении (todo апдейт стейта вынести?)
 
-    gameEngine.ts - отвечает за состояние игры в зависимости от глобального стейта: загрузка обложки игры до старта, завершение игры, завершение уровня, конец игры, старт, пауза.
+    gameEngine.ts - отвечает за главный цикл анимации игры, его запуск, сброс и остановка. В цикле проверяется актуальность обьектов умер/жив корабль, виден/спрятан выстрел, умер ли плейер, а также закончился ли уровень. Происходит апдейт стейта всех обьектов (кораблей, выстрелов). Вызывается gamePainter для отрисовки нужного обьекта. Проверяются коллизии. Выставляется нужный стейт игры в глобальном хранилище, в зависимости от глобального стейта загружается нужный экран (обложки игры до старта, завершение игры, завершение уровня, конец игры, старт, пауза)
 
-    gameAnimator.ts - отвечает за главный цикл анимации игры, его старт и пауза. Главный цикл умеет отрисовывать корабли и выстрелы из стейта игры. В цикле проходят проверки (умер плейер или уровень закончился) и выставляется стейт игры
+    gamePainter.ts - умеет отрисовывать базовые обьекты с картинкой, координатами и размерами, умеет рисовать бекграунд, экран начала игры и экран
 
 game.module.scss - стили для компонента игры
 game.tsx - компонент игры
