+++ packages/client/src/app/pages/game/gameEngine/gameControlManager.ts
@@ -1,0 +1,41 @@
+import { GameShot } from './types/gameTypes';
+import { ShotType } from './types/commonTypes';
+import gameState from './store/gameState';
+
+// todo move it in some control module ?
+const ControlKeys = {
+    LEFT: 'ArrowLeft',
+    UP: 'ArrowUp',
+    RIGHT: 'ArrowRight',
+    DOWN: 'ArrowDown',
+    PAUSE: 'Enter',
+    SHOOT: 'a',
+};
+
+export type TDirection = 'Up' | 'Down' | 'Left' | 'Right';
+
+export class GameControlManager {
+    public static gameControlPressed = (event: KeyboardEvent, time: number) => {
+        let direction: TDirection | undefined;
+        if (event.key === ControlKeys.UP) {
+            direction = 'Up';
+        } else if (event.key === ControlKeys.DOWN) {
+            direction = 'Down';
+        } else if (event.key === ControlKeys.LEFT) {
+            direction = 'Left';
+        } else if (event.key === ControlKeys.RIGHT) {
+            direction = 'Right';
+        }
+        const { player } = gameState;
+        if (direction) {
+            player.updateState(false, direction);
+        }
+
+        if (event.key === ControlKeys.SHOOT) {
+            console.log(event.key);
+            console.log('add shot');
+            const coordinates = player.getState().getCoordinates();
+            gameState.shots.push(new GameShot(ShotType.Player, coordinates, time));
+        }
+    };
+}
