+++ packages/client/src/app/pages/startGame/startGame.tsx
@@ -1,19 +1,17 @@
 /* eslint-disable no-plusplus */
 import React, { FC, useEffect, useState } from 'react';
-import { useDispatch } from 'react-redux';
 import Progressbar from '@/app/components/common/progressbar/progressbar';
 import style from './startgame.module.scss';
-import { setGameState } from '@/app/store/slices/gameSlice';
 import { GlobalGameState } from '../game/gameEngine/types/objectState';
+import GameEngine from '../game/gameEngine/gameEngine';
 
 const StartGame: FC = () => {
     const [value, setValue] = useState(0);
-    const dispatch = useDispatch();
     useEffect(() => {
         const interval = window.setInterval(() => {
             if (value >= 100) {
                 clearInterval(interval);
-                dispatch(setGameState(GlobalGameState.LevelStarted));
+                GameEngine.getInstance().setGameState(GlobalGameState.LevelStarted);
                 return;
             }
             setValue(value + 1);
